+++ R/decomp.design.R
@@ -145,8 +145,7 @@ decomp.design <- function(x, tau.preset = x$tau.preset, warn = gs("warn"),
   
   
   chkclass(x, "netmeta")
-  chksuitable(x, "Decomposition of designs",
-              classes = c("netmeta.crossnma", "netmeta.multinma"))
+  chksuitable(x, "Decomposition of designs", classes = gs(".other_nma"))
   ##
   if (x$n == 2) {
     warning("No decomposition possible for network meta-analysis ",