+++ R/netmeta-package.R
@@ -291,7 +291,7 @@
 #'   unit viewport
 #' 
 #' @importFrom dplyr %>% filter select rename starts_with relocate last_col
-#'   mutate if_else bind_rows
+#'   mutate if_else bind_rows pull
 #' 
 #' @importFrom magrittr %<>%