+++ R/rankogram.netmeta.R
@@ -138,9 +138,10 @@ rankogram.netmeta <- function(x, nsim = gs("nsim"),
     deprecated(random, missing(random), args, "comb.random", warn.deprecated)
   chklogical(random)
   #
-  # Additional checks for crossnma and multinma objects (and return results)
+  # Additional checks for crossnma, gemtc and multinma objects
+  # (and return results)
   #
-  if (inherits(x, c("netmeta.crossnma", "netmeta.multinma"))) {
+  if (inherits(x, gs(".other_nma"))) {
     if (!x$keep.samples)
       stop("Input for argument 'x' is a netmeta.", x$method,
            " object without samples; recreate ", x$method,